Move business logic into `transform.js` file
