+++ package.json
@@ -26,7 +26,7 @@
   },
   "jest": {
     "collectCoverageFrom": [
-      "transforms/angle-brackets/index.js"
+      "transforms/angle-brackets/transform.js"
     ]
   },
   "dependencies": {
